Tests: real timings and real speech in the fixtures
The blocked labeling vectors carried no timing data at all — the
harnesses synthesised a uniform 1 s stride — so gap clustering could not
be exercised by them and they encoded the abandoned algorithm's answers.
They now ship explicit start/end times taken from a real recording, and
the basic vector's expected labels match the original tool's output on
that slice. The vector file remains a Python/TypeScript parity contract,
not a claim about which labels are correct.

Detection fixtures use the bundled demo clip rather than sine bursts:
Silero is trained on speech and correctly ignores tones, so synthetic
audio was only ever testable against the detector we removed.

+++ tests/test_audio_io.py
@@ -20,7 +20,7 @@ from clicketysplit.audio_io import (
 
 
 def _sine(duration_sec: float, freq_hz: float, sr: int) -> np.ndarray:
-    t = np.arange(int(round(duration_sec * sr))) / sr
+    t = np.arange(round(duration_sec * sr)) / sr
     return (0.5 * np.sin(2.0 * np.pi * freq_hz * t)).astype(np.float32)
 
 
@@ -92,7 +92,7 @@ def test_concatenate_two_files(tmp_path: Path) -> None:
     assert result.audio.ndim == 1
 
     expected_total_sec = 0.3 + 0.1 + 0.3
-    assert result.audio.shape[0] == int(round(expected_total_sec * sr))
+    assert result.audio.shape[0] == round(expected_total_sec * sr)
 
     assert len(result.boundaries) == 2
     b0, b1 = result.boundaries
